PythonChallenge/requirements.txt
Pillow==3.4.1
requests==2.11.1
